chatbot/requirements.txt
anyio==4.4.0
certifi==2024.8.30
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
ollama==0.3.3
psycopg2==2.9.9
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
